M0: export direct Schedule service contract
